requirements/requirements-setup.txt
# Any modifications here must also be done in the pyproject.toml file
setuptools >= 42
wheel
cython >= 3.0.0a9
versioneer-518
numpy
